requirements.txt
annotated-types==0.7.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
defusedxml==0.7.1
diskcache==5.6.3
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
filelock==3.18.0
fonttools==4.58.5
fpdf2==2.8.3
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
llama_cpp_python==0.3.12
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
numpy==2.3.1
packaging==25.0
pfzy==0.3.4
pillow==11.3.0
preshed==3.0.10
prompt_toolkit==3.0.51
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.3.0.post1
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
torch==2.7.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
wheel==0.45.1
wrapt==1.17.2
